+++ port/posix/lisle/sys/thread.h
@@ -41,7 +41,7 @@ struct thread
 	lisle::mutex guard;
 	~thread ();
 	thread ();
-	uint32_t id () const;
+	uint64_t id () const;
 	bool equal (const thread& that) const;
 	bool equal (const thread* that) const;
 };
